refactor(materializer): extract interaction-patch conversation strategies (file #3)
Lift the 6 interaction-patch conversation fast paths into
interaction-patch-conversations.ts; dispatcher imports them. Net 74/74, check 0.
